Fix txUtils invalid `.length` on function
The property `getCommandsList` is a function that returns an `Array`. The `.length` was probably intended for the result of calling the function not the function itself.  Calling `.length` on a function returns the number of arguments in the function signature which is probably not what the author intended to use for  `addCommands`.